refactor(context): decompose ContextManager to <=300 lines via mixins (#101)
manager.py was 878 lines of thin delegating method stubs + docstrings (the
pipeline logic already lived in build.py / route_build.py / call_prompt.py /
ingest.py). Move the method surface into flat, single-level partial-class
mixins so the orchestrator meets the repo's <=300-line module guideline
without dropping any public method:

- _ManagerState (context/_manager_base.py): private-attribute + _build
  contract the mixins inherit and the delegate pipeline modules type their
  `manager` parameter against. ContextManager is a structural subtype, so
  every existing call site is unchanged.
- _IngestMixin / _BuildMixin / _RoutingMixin: ingestion+facts+episodes,
  core build, and route/call-prompt method surfaces respectively.
- manager.py is now 239 lines (init, properties, drilldown, composition);
  every module is <=300.

No public API change: all 21 methods stay on ContextManager. Full suite
(1927 passed) and `make ci` pass unmodified.
