+++ src/contextweaver/context/manager.py
@@ -10,21 +10,27 @@
 6. :func:`~contextweaver.context.dedup.deduplicate_candidates` — near-duplicate removal
 7. :func:`~contextweaver.context.selection.select_and_pack` — budget-aware selection
 8. :func:`~contextweaver.context.prompt.render_context` — prompt assembly
+
+The class is a thin orchestrator: its public methods are split across small
+*partial-class* mixins (issue #101) to keep this module within the project's
+<=300 lines per module guideline.  The mixins (:class:`_IngestMixin`,
+:class:`_BuildMixin`, :class:`_RoutingMixin`) are flat, single-level, and not
+part of the public API; the heavy logic they delegate to lives in
+:mod:`contextweaver.context.ingest`, :mod:`~contextweaver.context.build`,
+:mod:`~contextweaver.context.route_build`, and
+:mod:`~contextweaver.context.call_prompt`.
 """
 
 from __future__ import annotations
 
-import logging
-from typing import TYPE_CHECKING, Any, Literal, overload
+from typing import Any
 
 from contextweaver.config import ContextBudget, ContextPolicy, ScoringConfig
 from contextweaver.context import ingest as _ingest
-from contextweaver.context import route_build as _route_build
-from contextweaver.context.build import run_build_pipeline
-from contextweaver.context.call_prompt import run_call_prompt_build
-from contextweaver.context.explanation import ContextBuildExplanation
+from contextweaver.context._manager_build import _BuildMixin
+from contextweaver.context._manager_ingest import _IngestMixin
+from contextweaver.context._manager_routing import _RoutingMixin
 from contextweaver.context.views import ViewRegistry
-from contextweaver.envelope import ContextPack, ResultEnvelope
 from contextweaver.metrics import MetricsCollector
 from contextweaver.profiles import Mode, ProfileConfig
 from contextweaver.protocols import (
@@ -41,21 +47,12 @@ from contextweaver.protocols import (
 )
 from contextweaver.store import StoreBundle
 from contextweaver.store.artifacts import InMemoryArtifactStore
-from contextweaver.store.episodic import Episode, InMemoryEpisodicStore
+from contextweaver.store.episodic import InMemoryEpisodicStore
 from contextweaver.store.event_log import InMemoryEventLog
-from contextweaver.store.facts import Fact, InMemoryFactStore
-from contextweaver.types import ContextItem, Phase
-
-if TYPE_CHECKING:
-    from contextweaver.envelope import ChoiceCard
-    from contextweaver.routing.catalog import Catalog
-    from contextweaver.routing.history import RouteHistory
-    from contextweaver.routing.router import Router, RouteResult
-
-logger = logging.getLogger("contextweaver.context")
+from contextweaver.store.facts import InMemoryFactStore
 
 
-class ContextManager:
+class ContextManager(_IngestMixin, _BuildMixin, _RoutingMixin):
     """Orchestrates the full context compilation pipeline.
 
     Args:
@@ -187,253 +184,6 @@ class ContextManager:
         return self._budget
 
     # ------------------------------------------------------------------
-    # Ingestion helpers
-    # ------------------------------------------------------------------
-
-    def ingest(self, item: ContextItem) -> None:
-        """Append *item* to the event log.
-
-        Args:
-            item: The context item to ingest.
-        """
-        _ingest.ingest_item(self._event_log, item)
-
-    def ingest_sync(self, item: ContextItem) -> None:
-        """Synchronous alias for :meth:`ingest`."""
-        self.ingest(item)
-
-    async def ingest_async(self, item: ContextItem) -> None:
-        """Async alias for :meth:`ingest`."""
-        self.ingest(item)
-
-    def ingest_envelope(
-        self,
-        tool_call_id: str,
-        envelope: ResultEnvelope,
-        tool_name: str = "",
-    ) -> ContextItem:
-        """Ingest an already-firewalled :class:`ResultEnvelope` (canonical path).
-
-        This is the **canonical, Frame-shaped ingestion path** for the Weaver
-        Stack (weaver-spec I-05).  The execution boundary (e.g. agent-kernel)
-        owns firewalling and hands contextweaver a :class:`ResultEnvelope` —
-        the native preimage of a weaver-spec ``Frame``.  contextweaver appends
-        a summary-only :class:`ContextItem` carrying the envelope's artifact
-        handle and performs budgeted selection/packing; it does **not**
-        re-derive firewalling from raw output.
-
-        weaver-spec users convert a ``Frame`` first::
-
-            from contextweaver.adapters.weaver_contracts import from_weaver_frame
-            mgr.ingest_envelope(tool_call_id, from_weaver_frame(frame))
-
-        See :doc:`the context-firewall boundary doc </context_firewall_boundary>`
-        for who firewalls what and where the seam sits.
-
-        Args:
-            tool_call_id: ID of the originating tool call.
-            envelope: An already-firewalled :class:`ResultEnvelope`.
-            tool_name: Human-readable tool name.
-
-        Returns:
-            The appended :class:`ContextItem` (summary text + artifact handle).
-        """
-        return _ingest.ingest_envelope(
-            event_log=self._event_log,
-            estimator=self._estimator,
-            tool_call_id=tool_call_id,
-            envelope=envelope,
-            tool_name=tool_name,
-        )
-
-    def ingest_envelope_sync(
-        self,
-        tool_call_id: str,
-        envelope: ResultEnvelope,
-        tool_name: str = "",
-    ) -> ContextItem:
-        """Synchronous alias for :meth:`ingest_envelope`."""
-        return self.ingest_envelope(tool_call_id, envelope, tool_name)
-
-    def ingest_tool_result(
-        self,
-        tool_call_id: str,
-        raw_output: str,
-        tool_name: str = "",
-        media_type: str = "text/plain",
-        firewall_threshold: int = 2000,
-    ) -> tuple[ContextItem, ResultEnvelope]:
-        """Ingest a *raw* tool result, running the context firewall locally.
-
-        .. note::
-
-            **Non-canonical for spec compliance (weaver-spec I-05).** This API
-            accepts raw output and re-derives firewalling inside contextweaver.
-            The canonical seam is :meth:`ingest_envelope`, where the execution
-            boundary firewalls and hands over a :class:`ResultEnvelope` /
-            ``Frame``.  Prefer :meth:`ingest_envelope` when integrating with an
-            agent-kernel-style execution boundary; this method remains for
-            standalone use where contextweaver owns the firewall.
-
-        If the raw output exceeds *firewall_threshold* characters it is stored
-        in the artifact store and the LLM sees only a summary.  Small outputs
-        are also stored in the artifact store (with ``artifact_ref`` set on the
-        returned item) to enable drilldown on all tool results regardless of
-        size.
-
-        Args:
-            tool_call_id: ID of the originating tool call.
-            raw_output: Raw tool output string.
-            tool_name: Human-readable tool name.
-            media_type: MIME type of the output.
-            firewall_threshold: Character threshold above which the firewall
-                stores the raw output out-of-band.
-
-        Returns:
-            A ``(ContextItem, ResultEnvelope)`` tuple.  The item always has a
-            non-``None`` ``artifact_ref``.
-        """
-        return _ingest.ingest_tool_result(
-            event_log=self._event_log,
-            artifact_store=self._artifact_store,
-            hook=self._hook,
-            view_registry=self._view_registry,
-            summarizer=self._summarizer,
-            extractor=self._extractor,
-            estimator=self._estimator,
-            tool_call_id=tool_call_id,
-            raw_output=raw_output,
-            tool_name=tool_name,
-            media_type=media_type,
-            firewall_threshold=firewall_threshold,
-        )
-
-    def ingest_tool_result_sync(
-        self,
-        tool_call_id: str,
-        raw_output: str,
-        tool_name: str = "",
-        media_type: str = "text/plain",
-        firewall_threshold: int = 2000,
-    ) -> tuple[ContextItem, ResultEnvelope]:
-        """Synchronous alias for :meth:`ingest_tool_result`."""
-        return self.ingest_tool_result(
-            tool_call_id, raw_output, tool_name, media_type, firewall_threshold
-        )
-
-    def ingest_mcp_result(
-        self,
-        tool_call_id: str,
-        mcp_result: dict[str, Any],
-        tool_name: str,
-        firewall_threshold: int = 2000,
-    ) -> tuple[ContextItem, ResultEnvelope]:
-        """Ingest a raw MCP tool result with full artifact persistence.
-
-        .. note::
-
-            **Non-canonical for spec compliance (weaver-spec I-05).** Like
-            :meth:`ingest_tool_result`, this accepts a raw upstream payload and
-            re-derives firewalling inside contextweaver.  When the execution
-            boundary already produces a :class:`ResultEnvelope` / ``Frame``, use
-            :meth:`ingest_envelope` instead.  This method remains the
-            happy-path for direct MCP integration where contextweaver owns the
-            firewall.
-
-        It:
-
-        1. Parses the MCP result via :func:`mcp_result_to_envelope`.
-        2. Stores binary artifacts (images, resources) in the artifact store.
-        3. Applies the context firewall for large text outputs.
-        4. Appends the resulting :class:`ContextItem` to the event log.
-
-        Args:
-            tool_call_id: ID of the originating tool call.
-            mcp_result: Raw MCP tool result dict (with ``content`` list).
-            tool_name: Human-readable tool name.
-            firewall_threshold: Character threshold above which text output
-                is stored out-of-band via the firewall.
-
-        Returns:
-            A ``(ContextItem, ResultEnvelope)`` tuple with all artifacts
-            persisted in the artifact store.
-        """
-        return _ingest.ingest_mcp_result(
-            event_log=self._event_log,
-            artifact_store=self._artifact_store,
-            hook=self._hook,
-            summarizer=self._summarizer,
-            extractor=self._extractor,
-            estimator=self._estimator,
-            tool_call_id=tool_call_id,
-            mcp_result=mcp_result,
-            tool_name=tool_name,
-            firewall_threshold=firewall_threshold,
-        )
-
-    def ingest_mcp_result_sync(
-        self,
-        tool_call_id: str,
-        mcp_result: dict[str, Any],
-        tool_name: str,
-        firewall_threshold: int = 2000,
-    ) -> tuple[ContextItem, ResultEnvelope]:
-        """Synchronous alias for :meth:`ingest_mcp_result`."""
-        return self.ingest_mcp_result(tool_call_id, mcp_result, tool_name, firewall_threshold)
-
-    def add_fact(self, key: str, value: str, metadata: dict[str, Any] | None = None) -> None:
-        """Store a fact in the fact store.
-
-        Args:
-            key: Fact key.
-            value: Fact value.
-            metadata: Optional metadata dict.
-        """
-        fact_id = f"fact:{key}:{len(self._fact_store.all())}"
-        self._fact_store.put(
-            Fact(
-                fact_id=fact_id,
-                key=key,
-                value=value,
-                metadata=metadata or {},
-            )
-        )
-
-    def add_fact_sync(self, key: str, value: str, metadata: dict[str, Any] | None = None) -> None:
-        """Synchronous alias for :meth:`add_fact`."""
-        self.add_fact(key, value, metadata)
-
-    def add_episode(
-        self,
-        episode_id: str,
-        summary: str,
-        metadata: dict[str, Any] | None = None,
-    ) -> None:
-        """Store an episodic memory summary.
-
-        Args:
-            episode_id: Unique episode identifier.
-            summary: Summary text.
-            metadata: Optional metadata dict.
-        """
-        self._episodic_store.add(
-            Episode(
-                episode_id=episode_id,
-                summary=summary,
-                metadata=metadata or {},
-            )
-        )
-
-    def add_episode_sync(
-        self,
-        episode_id: str,
-        summary: str,
-        metadata: dict[str, Any] | None = None,
-    ) -> None:
-        """Synchronous alias for :meth:`add_episode`."""
-        self.add_episode(episode_id, summary, metadata)
-
-    # ------------------------------------------------------------------
     # Drilldown
     # ------------------------------------------------------------------
 
@@ -487,392 +237,3 @@ class ContextManager:
     ) -> str:
         """Synchronous alias for :meth:`drilldown`."""
         return self.drilldown(handle, selector, inject=inject, parent_id=parent_id)
-
-    # ------------------------------------------------------------------
-    # Core pipeline
-    # ------------------------------------------------------------------
-
-    def _build(
-        self,
-        phase: Phase = Phase.answer,
-        query: str = "",
-        query_tags: list[str] | None = None,
-        header: str = "",
-        footer: str = "",
-        budget_tokens: int | None = None,
-        hints: list[str] | None = None,
-        extra: dict[str, Any] | None = None,
-        explain: bool = False,
-    ) -> tuple[ContextPack, ContextBuildExplanation | None]:
-        """Run the full context compilation pipeline (synchronous core).
-
-        All eight pipeline steps are pure computation, so no ``await`` is
-        needed.  Both :meth:`build` (async) and :meth:`build_sync` delegate
-        here.
-
-        Args:
-            phase: Active execution phase.
-            query: User query string used for relevance scoring.
-            query_tags: Optional tag list to boost tag-matched items.
-            header: Optional prompt header text.
-            footer: Optional prompt footer text.
-            budget_tokens: Override the default phase budget.
-            hints: Additional hint tags for scoring.
-            extra: Reserved for future pipeline extensions.
-            explain: When ``True``, collect explanation-specific
-                intermediate state and return a populated
-                :class:`ContextBuildExplanation`.  When ``False``
-                (default), skip explanation overhead and return
-                ``None`` as the second tuple element (issue #291).
-
-        Returns:
-            A 2-tuple ``(pack, explanation)``.  *explanation* is
-            ``None`` when *explain* is ``False``.
-        """
-        return run_build_pipeline(
-            self,
-            phase=phase,
-            query=query,
-            query_tags=query_tags,
-            header=header,
-            footer=footer,
-            budget_tokens=budget_tokens,
-            hints=hints,
-            extra=extra,
-            explain=explain,
-        )
-
-    @overload
-    async def build(
-        self,
-        phase: Phase = ...,
-        query: str = ...,
-        query_tags: list[str] | None = ...,
-        header: str = ...,
-        footer: str = ...,
-        budget_tokens: int | None = ...,
-        hints: list[str] | None = ...,
-        extra: dict[str, Any] | None = ...,
-        *,
-        explain: Literal[False] = False,
-    ) -> ContextPack: ...
-
-    @overload
-    async def build(
-        self,
-        phase: Phase = ...,
-        query: str = ...,
-        query_tags: list[str] | None = ...,
-        header: str = ...,
-        footer: str = ...,
-        budget_tokens: int | None = ...,
-        hints: list[str] | None = ...,
-        extra: dict[str, Any] | None = ...,
-        *,
-        explain: Literal[True],
-    ) -> tuple[ContextPack, ContextBuildExplanation]: ...
-
-    async def build(
-        self,
-        phase: Phase = Phase.answer,
-        query: str = "",
-        query_tags: list[str] | None = None,
-        header: str = "",
-        footer: str = "",
-        budget_tokens: int | None = None,
-        hints: list[str] | None = None,
-        extra: dict[str, Any] | None = None,
-        *,
-        explain: bool = False,
-    ) -> ContextPack | tuple[ContextPack, ContextBuildExplanation]:
-        """Asynchronously compile a :class:`~contextweaver.envelope.ContextPack`.
-
-        The current pipeline is fully synchronous; this ``async`` wrapper
-        exists so callers can ``await`` it today and benefit from true
-        async I/O if the pipeline gains ``await``-able steps in the future.
-
-        See :meth:`_build` for full parameter documentation.
-
-        Args:
-            phase: Active execution phase.
-            query: User query string used for relevance scoring.
-            query_tags: Optional tag list to boost tag-matched items.
-            header: Optional prompt header text.
-            footer: Optional prompt footer text.
-            budget_tokens: Override the default phase budget.
-            hints: Additional hint tags for scoring.
-            extra: Reserved for future pipeline extensions.
-            explain: Keyword-only.  When ``True``, the method returns a
-                ``(pack, explanation)`` tuple where *explanation* is a
-                :class:`ContextBuildExplanation` capturing per-candidate
-                scoring + drop reasons.  Default ``False`` preserves the
-                bare :class:`ContextPack` return shape (issue #291).
-
-        Returns:
-            A :class:`~contextweaver.envelope.ContextPack` ready for
-            the LLM, or a ``(pack, explanation)`` tuple when
-            ``explain=True``.
-        """
-        pack, explanation = self._build(
-            phase=phase,
-            query=query,
-            query_tags=query_tags,
-            header=header,
-            footer=footer,
-            budget_tokens=budget_tokens,
-            hints=hints,
-            extra=extra,
-            explain=explain,
-        )
-        if explain:
-            assert explanation is not None
-            return (pack, explanation)
-        return pack
-
-    @overload
-    def build_sync(
-        self,
-        phase: Phase = ...,
-        query: str = ...,
-        query_tags: list[str] | None = ...,
-        header: str = ...,
-        footer: str = ...,
-        budget_tokens: int | None = ...,
-        hints: list[str] | None = ...,
-        extra: dict[str, Any] | None = ...,
-        *,
-        explain: Literal[False] = False,
-    ) -> ContextPack: ...
-
-    @overload
-    def build_sync(
-        self,
-        phase: Phase = ...,
-        query: str = ...,
-        query_tags: list[str] | None = ...,
-        header: str = ...,
-        footer: str = ...,
-        budget_tokens: int | None = ...,
-        hints: list[str] | None = ...,
-        extra: dict[str, Any] | None = ...,
-        *,
-        explain: Literal[True],
-    ) -> tuple[ContextPack, ContextBuildExplanation]: ...
-
-    def build_sync(
-        self,
-        phase: Phase = Phase.answer,
-        query: str = "",
-        query_tags: list[str] | None = None,
-        header: str = "",
-        footer: str = "",
-        budget_tokens: int | None = None,
-        hints: list[str] | None = None,
-        extra: dict[str, Any] | None = None,
-        *,
-        explain: bool = False,
-    ) -> ContextPack | tuple[ContextPack, ContextBuildExplanation]:
-        """Synchronous entry point — delegates to :meth:`_build`.
-
-        Works inside Jupyter notebooks, FastAPI handlers, and any other
-        environment where an event loop is already running.
-
-        See :meth:`_build` for full parameter documentation.
-
-        Args:
-            phase: Active execution phase.
-            query: User query string used for relevance scoring.
-            query_tags: Optional tag list to boost tag-matched items.
-            header: Optional prompt header text.
-            footer: Optional prompt footer text.
-            budget_tokens: Override the default phase budget.
-            hints: Additional hint tags for scoring.
-            extra: Reserved for future pipeline extensions.
-            explain: Keyword-only.  Same semantics as
-                :meth:`build` (issue #291).
-
-        Returns:
-            A :class:`~contextweaver.envelope.ContextPack` ready for
-            the LLM, or a ``(pack, explanation)`` tuple when
-            ``explain=True``.
-        """
-        pack, explanation = self._build(
-            phase=phase,
-            query=query,
-            query_tags=query_tags,
-            header=header,
-            footer=footer,
-            budget_tokens=budget_tokens,
-            hints=hints,
-            extra=extra,
-            explain=explain,
-        )
-        if explain:
-            assert explanation is not None
-            return (pack, explanation)
-        return pack
-
-    # ------------------------------------------------------------------
-    # Route-integrated build
-    # ------------------------------------------------------------------
-
-    def build_route_prompt(
-        self,
-        goal: str,
-        query: str,
-        router: Router,
-        budget_tokens: int | None = None,
-        *,
-        history: RouteHistory | None = None,
-        history_from_log: bool = True,
-    ) -> tuple[ContextPack, list[ChoiceCard], RouteResult]:
-        """Route, build context, and assemble a prompt with choice cards.
-
-        Runs the router to find the best tools for *query*, then builds a
-        :class:`ContextPack` for the ``route`` phase with choice cards
-        appended.
-
-        Args:
-            goal: High-level goal description.
-            query: User query string.
-            router: The :class:`Router` to use for tool routing.
-            budget_tokens: Optional budget override.
-            history: Optional pre-built :class:`RouteHistory` (issue #27).
-                When ``None`` and *history_from_log* is ``True``, the
-                manager auto-constructs one from its event log.
-            history_from_log: When ``True`` (default) and *history* is
-                ``None``, build a :class:`RouteHistory` from the event log
-                — already-called tools are deprioritised on subsequent
-                routing calls in the same session.  Set to ``False`` to
-                preserve pre-#27 stateless routing.
-
-        Returns:
-            A 3-tuple ``(pack, cards, route_result)``.
-        """
-        return _route_build.build_route_prompt(
-            self,
-            goal,
-            query,
-            router,
-            budget_tokens,
-            history=history,
-            history_from_log=history_from_log,
-        )
-
-    def build_route_prompt_sync(
-        self,
-        goal: str,
-        query: str,
-        router: Router,
-        budget_tokens: int | None = None,
-        *,
-        history: RouteHistory | None = None,
-        history_from_log: bool = True,
-    ) -> tuple[ContextPack, list[ChoiceCard], RouteResult]:
-        """Synchronous alias for :meth:`build_route_prompt`."""
-        return self.build_route_prompt(
-            goal,
-            query,
-            router,
-            budget_tokens,
-            history=history,
-            history_from_log=history_from_log,
-        )
-
-    # ------------------------------------------------------------------
-    # Call-phase prompt (schema injection)
-    # ------------------------------------------------------------------
-
-    def _build_call_prompt(
-        self,
-        tool_id: str,
-        query: str,
-        catalog: Catalog,
-        schema: dict[str, Any] | None = None,
-        examples: list[str] | None = None,
-        constraints: dict[str, Any] | None = None,
-        budget_tokens: int | None = None,
-    ) -> ContextPack:
-        """Build a ``Phase.call`` prompt with the tool's full schema injected.
-
-        Hydrates the selected tool from *catalog*, delegates header assembly
-        to :func:`~contextweaver.context.call_prompt.build_schema_header`,
-        then runs the standard context pipeline with the call-phase budget.
-
-        Args:
-            tool_id: ID of the tool selected during routing.
-            query: User query string for relevance scoring.
-            catalog: The :class:`Catalog` containing *tool_id*.
-            schema: Override schema dict (replaces hydrated ``args_schema``
-                in the prompt; hydration still occurs for item metadata).
-            examples: Override example strings (replaces hydrated examples
-                in the prompt; hydration still occurs for item metadata).
-            constraints: Override constraints dict (replaces hydrated
-                ``constraints`` in the prompt; hydration still occurs for
-                item metadata).
-            budget_tokens: Override the default ``Phase.call`` budget.
-
-        Returns:
-            A :class:`~contextweaver.envelope.ContextPack` for ``Phase.call``.
-
-        Raises:
-            ItemNotFoundError: If *tool_id* is not in *catalog*.
-        """
-        return run_call_prompt_build(
-            self,
-            tool_id=tool_id,
-            query=query,
-            catalog=catalog,
-            schema=schema,
-            examples=examples,
-            constraints=constraints,
-            budget_tokens=budget_tokens,
-        )
-
-    async def build_call_prompt(
-        self,
-        tool_id: str,
-        query: str,
-        catalog: Catalog,
-        schema: dict[str, Any] | None = None,
-        examples: list[str] | None = None,
-        constraints: dict[str, Any] | None = None,
-        budget_tokens: int | None = None,
-    ) -> ContextPack:
-        """Async wrapper for :meth:`_build_call_prompt`.
-
-        See :meth:`_build_call_prompt` for parameter documentation.
-        """
-        return self._build_call_prompt(
-            tool_id=tool_id,
-            query=query,
-            catalog=catalog,
-            schema=schema,
-            examples=examples,
-            constraints=constraints,
-            budget_tokens=budget_tokens,
-        )
-
-    def build_call_prompt_sync(
-        self,
-        tool_id: str,
-        query: str,
-        catalog: Catalog,
-        schema: dict[str, Any] | None = None,
-        examples: list[str] | None = None,
-        constraints: dict[str, Any] | None = None,
-        budget_tokens: int | None = None,
-    ) -> ContextPack:
-        """Synchronous alias for :meth:`_build_call_prompt`.
-
-        See :meth:`_build_call_prompt` for parameter documentation.
-        """
-        return self._build_call_prompt(
-            tool_id=tool_id,
-            query=query,
-            catalog=catalog,
-            schema=schema,
-            examples=examples,
-            constraints=constraints,
-            budget_tokens=budget_tokens,
-        )
